add super class
